Replace set-favicon with generate-fav util
